Flyt alle imports til toppen af sitelogger.py
Det er god Python-skik at putte imports øverst i koden. Det er også god
skik at dele sine imports op i tre sektioner:

1. imports fra standardbiblioteket (fx os, re & datetime)
2. imports fra tredjepart (fx numpy eller matplotlib)
3. imports fra "sig selv" (fx antenna eller receiver)

I dette tilfælde har vi kun 1 og 3, så her er der kun to sektioner af
imports. Sektionerne adskilles bare med en linje, det er ikke nødvendigt
med en forklarende kommentar - en dygtig Pythonudvikler ved godt hvad
det betyder.

For at kunne gøre ovenstående var det nødvendigt at omorganisere koden
lidt. Ellers ville der opstå problemer med at man importerer moduler,
der skal bruge Section, men Section defineres først efter imports. Det
kan man naturligvis ikke, så jeg har flyttet Section, SubSection og
SectionList over i sitelog/sections.py og tilpasset diverse imports så
de afspejler dette. Denne ændre skaber et problem med
_determine_line_type() som nu skal bruges i to forskellige moduler
(sitelogger.py og sections.py). Derfor er den flyttet til
sitelog/__init__.py og de to andre moduler importerer nu funktionen fra
denne.

NB: Bemærk at sitelog/__init__.py både starter og slutter med imports.
Det strider mod hvad jeg skriver øverst. __init__.py er altid lidt en
skævert i Pythonpakker. Det er ofte her man gemmer alle de beskidte
tricks, der gør det let at bruge en pakke. Lige i __init__py er den
slags unoder tilladt :-)
